dio_cloud_native/DIOServeless/FnFila/requirements.txt
azure-servicebus
python-dotenv
azure